requirements.txt
requests==2.18.4
pytesseract==0.2.0
lxml==4.1.1
Pillow==5.0.0
